python-ai/requirements.txt
fastapi
uvicorn[standard]
openai
grpcio
grpcio-tools
python-dotenv